+++ conan/tools/cmake/toolchain.py
@@ -10,8 +10,7 @@ from jinja2 import Template
 from conan.tools import CONAN_TOOLCHAIN_ARGS_FILE
 from conan.tools._compilers import architecture_flag, use_win_mingw
 from conan.tools.cmake.utils import is_multi_configuration, get_file_name
-from conan.tools.microsoft.toolchain import write_conanvcvars
-from conan.tools.microsoft.visual import vs_ide_version
+from conan.tools.microsoft.visual import vs_ide_version, write_conanvcvars
 from conans.errors import ConanException
 from conans.util.files import load, save
