+++ conan/tools/microsoft/__init__.py
@@ -1,5 +1,5 @@
 from conan.tools.microsoft.toolchain import MSBuildToolchain
 from conan.tools.microsoft.msbuild import MSBuild
 from conan.tools.microsoft.msbuilddeps import MSBuildDeps
-from conan.tools.microsoft.visual import msvc_runtime_flag
+from conan.tools.microsoft.visual import msvc_runtime_flag, VCVars
 from conan.tools.microsoft.subsystems import unix_path
