+++ conan/tools/microsoft/visual.py
@@ -1,7 +1,66 @@
 import os
+import textwrap
 
+from conan.tools.env.environment import register_environment_script
+from conans.client.tools import intel_compilervars_command
 from conans.client.tools.win import vs_installation_path
 from conans.errors import ConanException
+from conans.util.files import save
+
+CONAN_VCVARS_FILE = "conanvcvars.bat"
+
+
+class VCVars:
+    def __init__(self, conanfile):
+        self._conanfile = conanfile
+
+    def generate(self, auto_activate=True):
+        write_conanvcvars(self._conanfile, auto_activate=auto_activate)
+
+
+def write_conanvcvars(conanfile, auto_activate=True):
+    # FIXME: Write a VCVars generator for the final user
+    """
+    write a conanvcvars.bat file with the good args from settings
+    """
+    os_ = conanfile.settings.get_safe("os")
+    if os_ != "Windows":
+        return
+
+    compiler = conanfile.settings.get_safe("compiler")
+    cvars = None
+    if compiler == "intel":
+        cvars = intel_compilervars_command(conanfile)
+    elif compiler == "Visual Studio" or compiler == "msvc":
+        vs_version = vs_ide_version(conanfile)
+        vcvarsarch = vcvars_arch(conanfile)
+        vcvars_ver = None
+        if compiler == "Visual Studio":
+            toolset = conanfile.settings.get_safe("compiler.toolset")
+            if toolset is not None:
+                vcvars_ver = {"v140": "14.0",
+                              "v141": "14.1",
+                              "v142": "14.2"}.get(toolset)
+        else:
+            # Code similar to CMakeToolchain toolset one
+            compiler_version = str(conanfile.settings.compiler.version)
+            version_components = compiler_version.split(".")
+            assert len(version_components) >= 2  # there is a 19.XX
+            minor = version_components[1]
+            # The equivalent of compiler 19.26 is toolset 14.26
+            vcvars_ver = "14.{}".format(minor)
+        cvars = vcvars_command(vs_version, architecture=vcvarsarch, platform_type=None,
+                               winsdk_version=None, vcvars_ver=vcvars_ver)
+    if cvars:
+        content = textwrap.dedent("""\
+            @echo off
+            {}
+            """.format(cvars))
+        path = os.path.join(conanfile.generators_folder, CONAN_VCVARS_FILE)
+        save(path, content)
+
+        if auto_activate:
+            register_environment_script(conanfile, path)
 
 
 def vs_ide_version(conanfile):
